# Allow some array functions to operate in-place using a peep-hole optimization