@mvorisek 2023-04-11
> Note that this is limited to CV's at the moment, and can't handle more complex scenarios like array or object assignments.

Thank you for looking into this. `$x[$i] = array_merge($x[$i], expr)` is quite common, can it be optimized as well?

Also does this work for `$x = [...$x, expr]` (array spread)?

And can `$x = array_merge(expr, $x)` (array merge, but prepend the new values) be optimized as well?

Also can the optimization be asserted by a test (with `zend_test` ext if needed)?