@nielsdos 2023-04-11
> > Note that this is limited to CV's at the moment, and can't handle more complex scenarios like array or object assignments.
> 
> Thank you for looking into this. `$x[$i] = array_merge($x[$i], expr)` is quite common, can it be optimized as well?
> 

Not at the moment, but maybe in the future. The main goal of this PR is to show the idea works and to implement a first version of the idea. It can always be extended on in the future if desired. In any case I think it's important to first gather some feedback on this before continuing on.

> Also does this work for `$x = [...$x, expr]` (array spread)?
> 

No, because this PR optimizes the array functions, and what you show here must be implemented in the VM. I can look into the feasibility to optimize this next week (I'm quite busy now). But no promises on this one.

> And can `$x = array_merge(expr, $x)` (array merge, but prepend the new values) be optimized as well?
> 

I don't think so tbh. The reason is that the indices of the array may change, so we'd have to move a whole bunch of memory anyway. At that point a copy is not much more expensive.

> Also can the optimization be asserted by a test (with `zend_test` ext if needed)?

Kinda hard to do reliably tbh. I'd have to interpose the array duplication function to be able to count the amount of invocations. Getting this reliable is nasty. I'll first hear what other maintainers think about this before making more progress on this.