@mvorisek 2023-04-12
> test

Having this tested is very important, as user must be allowed to safely rely on the optimization. What about storing some flag, via `zend_test` into the source zval and testing the present/non-presence in the result?