@dstogov 2023-04-12
The benchmark results looks promising, but implementation is hacky.
I made only a quick look. I'll try to find more time to think how to implement this in a more general and better way.
Unfortunately, I hardly ever will have time before Monday.

Please, don't merge this yet.