@nielsdos 2023-04-23
Okay thanks for reviewing, I made sure the optimization only applies to RC1 inputs and I changed the name for the helper.