@mvorisek 2023-04-24
@dstogov as discussed in https://github.com/php/php-src/pull/11060#issuecomment-1504865831, do you think this can be tested easily with some help from `zend_test`?