@mvorisek 2023-04-24
> > @dstogov as discussed in [#11060 (comment)](https://github.com/php/php-src/pull/11060#issuecomment-1504865831), do you think this can be tested easily with some help from `zend_test`?
> 
> We don't need `zend_test` to test this. `usort($a1 + $a2, "cmp");` or something similar is enough. I think, this should be already covered.

@dstogov I meant the in-place performance:

> === array_merge $a = array_merge($a, ...) ===
before 4.3821 sec
after 0.0022 sec