@dstogov 2023-04-24
We don't prove performance with tests. It's possible to provide ``array_micro_bench.php`` similar to ``micro_bench.php``.