@mvorisek 2023-04-24
I would really like the optimization to be tested as the performance difference is dramatic. Do you think it is possible to save some flag in the original zval and test (after array_merge is called) if the zval flag is present in the assigned variable (if the flag is present, in-place was used). Or simply save `zval` C ptr before/after and compare it.