@nielsdos 2023-04-24
Thanks for reviewing I'll apply your suggestion.

@mvorisek Note that the latest version of the patch only performs the optimisation for temporaries. I dropped the `$variable = f($variable)` optimization because it had complications and wasn't safe.
I'm working on a new approach for the variable in-place modification in my fork. That new approach needs this PR first. The new approach also generalises to userland functions.
I think I can make a bench test file on my fork, it may make sense to do that there.